chore(chat): extract banner state hook and BannerRow component
